chore: fix linter to ignore eip3860 (#1)
* chore: fix linter to ignore eip3860

* chore: remove eip3860 and add as flag instead

* chore: run forge fmt

* chore: add test logic to pull from metadata or skip

* chore: update README lines

+++ src/utils/ChainFamilies.sol
@@ -14,46 +14,46 @@ library ChainFamilies {
     function getAllMainnets(LZProtocol protocol) internal view returns (uint32[] memory) {
         return _filterChains(protocol, "mainnet");
     }
-    
+
     function getAllSandboxes(LZProtocol protocol) internal view returns (uint32[] memory) {
         return _filterChains(protocol, "sandbox");
     }
 
     function _filterChains(LZProtocol protocol, string memory suffix) private view returns (uint32[] memory) {
         uint32[] memory allEids = protocol.getSupportedEids();
-        
+
         // First pass: count matches
         uint256 count = 0;
-        for (uint i = 0; i < allEids.length; i++) {
+        for (uint256 i = 0; i < allEids.length; i++) {
             ILZProtocol.ProtocolAddresses memory addr = protocol.getProtocolAddresses(allEids[i]);
             if (_contains(addr.chainName, suffix)) {
                 count++;
             }
         }
-        
+
         // Second pass: fill array
         uint32[] memory filtered = new uint32[](count);
         uint256 idx = 0;
-        for (uint i = 0; i < allEids.length; i++) {
+        for (uint256 i = 0; i < allEids.length; i++) {
             ILZProtocol.ProtocolAddresses memory addr = protocol.getProtocolAddresses(allEids[i]);
             if (_contains(addr.chainName, suffix)) {
                 filtered[idx] = allEids[i];
                 idx++;
             }
         }
-        
+
         return filtered;
     }
-    
+
     /// @dev Check if a string contains a substring (internal use only)
     function _contains(string memory haystack, string memory needle) private pure returns (bool) {
         bytes memory h = bytes(haystack);
         bytes memory n = bytes(needle);
         if (h.length < n.length) return false;
-        
-        for (uint i = 0; i <= h.length - n.length; i++) {
+
+        for (uint256 i = 0; i <= h.length - n.length; i++) {
             bool found = true;
-            for (uint j = 0; j < n.length; j++) {
+            for (uint256 j = 0; j < n.length; j++) {
                 if (h[i + j] != n[j]) {
                     found = false;
                     break;
